requirements.txt
selenium==4.1.5
dynaconf==3.1.8
ConcurrentLogHandler==0.9.1
psutil==5.9.0
python-dotenv==0.20.0
lxml==4.9.1
Werkzeug==2.2.2
marshmallow==3.21.3
ipdb
selenium-wire
webdriver_manager
blinker==1.7.0
mailjet_rest
requests
